+++ note.h
@@ -4,17 +4,17 @@
 class Note {
 private:
 	double Id;
-	size_t Balance;
+	double Balance;
 
 public:
 	Note(void);
 	Note(double id);
-	Note(double id, size_t summ);
+	Note(double id, double summ);
 	~Note(void);
 
-	void Push(size_t summ);
-	bool Pop(size_t summ);
-	size_t GetBalance(void);
+	void Push(double summ);
+	bool Pop(double summ);
+	double GetBalance(void);
 
 	Note& operator =(Note& note);
 	friend bool operator ==(Note& note1, Note& note2);
